requirements/dev.txt
-r common.txt
pytest==6.2.*
djangorestframework==3.12.*
pytest-django==4.4.*
